Move govuk_navigation_helpers into proper namespace
Follows the merging in of govuk_navigation_helpers (#244).

https://trello.com/c/WJlqy3VD
