--- lib/govuk_navigation_helpers.rb
@@ -1,10 +1,0 @@
-require 'active_support'
-require 'active_support/core_ext/hash/keys'
-require 'active_support/core_ext/object/blank'
-
-require_relative "govuk_navigation_helpers/breadcrumbs"
-require_relative "govuk_navigation_helpers/taxon_breadcrumbs"
-require_relative "govuk_navigation_helpers/taxonomy_sidebar"
-require_relative "govuk_navigation_helpers/rummager_taxonomy_sidebar_links"
-require_relative "govuk_navigation_helpers/curated_taxonomy_sidebar_links"
-require_relative "govuk_navigation_helpers/content_item"
