--- lib/govuk_navigation_helpers/services.rb
@@ -1,9 +1,0 @@
-require 'gds_api/rummager'
-
-module GovukNavigationHelpers
-  module Services
-    def self.rummager
-      @rummager ||= GdsApi::Rummager.new(Plek.find('rummager'))
-    end
-  end
-end
