+++ lib/govuk_publishing_components.rb
@@ -5,6 +5,13 @@ require "govuk_publishing_components/presenters/navigation_type"
 require "govuk_publishing_components/presenters/related_navigation_helper"
 require "govuk_publishing_components/presenters/step_by_step_nav_helper"
 require "govuk_publishing_components/presenters/step_nav_helper"
+require "govuk_publishing_components/presenters/breadcrumbs"
+require "govuk_publishing_components/presenters/taxon_breadcrumbs"
+require "govuk_publishing_components/presenters/services"
+require "govuk_publishing_components/presenters/taxonomy_sidebar"
+require "govuk_publishing_components/presenters/rummager_taxonomy_sidebar_links"
+require "govuk_publishing_components/presenters/curated_taxonomy_sidebar_links"
+require "govuk_publishing_components/presenters/content_item"
 
 require "govuk_publishing_components/app_helpers/taxon_breadcrumbs"
