+++ lib/govuk_publishing_components/presenters/contextual_navigation.rb
@@ -1,5 +1,3 @@
-require 'govuk_navigation_helpers'
-
 module GovukPublishingComponents
   module Presenters
     # @private
@@ -14,11 +12,11 @@ module GovukPublishingComponents
       end
 
       def taxonomy_sidebar
-        @taxonomy_sidebar ||= GovukNavigationHelpers::TaxonomySidebar.new(content_item).sidebar
+        @taxonomy_sidebar ||= TaxonomySidebar.new(content_item).sidebar
       end
 
       def taxon_breadcrumbs
-        @taxon_breadcrumbs ||= GovukNavigationHelpers::TaxonBreadcrumbs.new(content_item).breadcrumbs
+        @taxon_breadcrumbs ||= TaxonBreadcrumbs.new(content_item).breadcrumbs
       end
 
       def breadcrumbs
@@ -37,17 +35,17 @@ module GovukPublishingComponents
             }
           ]
         else
-          GovukNavigationHelpers::Breadcrumbs.new(content_item).breadcrumbs[:breadcrumbs]
+          Breadcrumbs.new(content_item).breadcrumbs[:breadcrumbs]
         end
       end
 
       def should_present_taxonomy_navigation?
-        navigation = GovukPublishingComponents::Presenters::NavigationType.new(content_item)
+        navigation = NavigationType.new(content_item)
         navigation.should_present_taxonomy_navigation?
       end
 
       def step_nav_helper
-        @step_nav_helper ||= GovukPublishingComponents::Presenters::StepNavHelper.new(content_item, request_path)
+        @step_nav_helper ||= StepNavHelper.new(content_item, request_path)
       end
     end
   end
